Type error fixes

+++ londonlaw/guiclient/MapWindow.py
@@ -173,34 +173,34 @@ class MapWindow(wx.ScrolledWindow):
       mapPixel     = locToPixel(loc, self.zoomLevel)
       stepX, stepY = self.GetScrollPixelsPerUnit()
       sX, sY       = self.GetViewStart()
-      self.labels[playerNum].Move(mapPixel[0] - 20 - sX*stepX, mapPixel[1] - 10 - sY*stepY)
+      self.labels[playerNum].Move(int(mapPixel[0] - 20 - sX*stepX), int(mapPixel[1] - 10 - sY*stepY))
 
 
    # remove a pushpin from the map bitmap
    def unDrawPushpin(self, playerNum):
       mapPixel = locToPixel(self.playerLoc[playerNum], self.zoomLevel)
       self.pushpinDC.SelectObject(self.pushpinBackgrounds[playerNum])
-      self.bmpDC.Blit(mapPixel[0]-self.pushpinOffset[0], mapPixel[1]-self.pushpinOffset[1], 
-            self.pushpinBackgrounds[playerNum].GetWidth(),
-            self.pushpinBackgrounds[playerNum].GetHeight(), self.pushpinDC, 0, 0)
+      self.bmpDC.Blit(int(mapPixel[0]-self.pushpinOffset[0]), int(mapPixel[1]-self.pushpinOffset[1]),
+            int(self.pushpinBackgrounds[playerNum].GetWidth()),
+            int(self.pushpinBackgrounds[playerNum].GetHeight()), self.pushpinDC, 0, 0)
 
 
    # update a pushpin background
    def updatePushpinBackground(self, playerNum):
       mapPixel = locToPixel(self.playerLoc[playerNum], self.zoomLevel)
       self.pushpinDC.SelectObject(self.pushpinBackgrounds[playerNum])
-      self.pushpinDC.Blit(0, 0, self.pushpinBackgrounds[playerNum].GetWidth(),
-            self.pushpinBackgrounds[playerNum].GetHeight(), 
-            self.bmpDC, mapPixel[0]-self.pushpinOffset[0], mapPixel[1]-self.pushpinOffset[1])
+      self.pushpinDC.Blit(0, 0, int(self.pushpinBackgrounds[playerNum].GetWidth()),
+            int(self.pushpinBackgrounds[playerNum].GetHeight()),
+            self.bmpDC, int(mapPixel[0]-self.pushpinOffset[0]), int(mapPixel[1]-self.pushpinOffset[1]))
 
 
    # blit a pushpin image to the map bitmap
    def drawPushpin(self, playerNum):
       mapPixel = locToPixel(self.playerLoc[playerNum], self.zoomLevel)
       self.pushpinDC.SelectObject(self.pushpins[playerNum])
-      self.bmpDC.Blit(mapPixel[0]-self.pushpinOffset[0], mapPixel[1]-self.pushpinOffset[1], 
-            self.pushpinBackgrounds[playerNum].GetWidth(),
-            self.pushpinBackgrounds[playerNum].GetHeight(), self.pushpinDC, 0, 0, wx.COPY, True)
+      self.bmpDC.Blit(int(mapPixel[0]-self.pushpinOffset[0]), int(mapPixel[1]-self.pushpinOffset[1]),
+            int(self.pushpinBackgrounds[playerNum].GetWidth()),
+            int(self.pushpinBackgrounds[playerNum].GetHeight()), self.pushpinDC, 0, 0, wx.COPY, True)
 
 
    # center the map on a particular player
